changed variable and declaration. only shifted attributes around but go over it again because it will be important for variable tracking
