+++ Algorithm-visualizer/AST/ast.hpp
@@ -59,8 +59,6 @@ virtual std::string get_type() = 0; // Statement or Expression   (this is an abs
 //and Expression and inherited by every other subclass below)
 virtual std::string get_subtype() = 0; // Block, Declaration, UnOp, BinOp, ...
 //(this is an abstract method which will be defined in subclasses Block, Declaration, ...)
-private:
-    std::map<char, char> variables;
 };
