+++ Algorithm-visualizer/AST/variable.hpp
@@ -18,7 +18,15 @@ class Variable : public Expression {
     void set_name(char n){name = n;};
     void set_type(var_type t){type = t;};
     std::string get_name(){return name;};
-    var_type get_var_type(){return type;};
+    std::string get_var_type(){
+        switch(type) {
+        case 0: return "unknown variable type";
+        case 1: return "string";
+        case 2: return "integer";
+        case 3: return "char";
+        case 4: return "floating_point";
+        }
+    }
     std::string get_subtype(){return "Variable";};
 
 private:
@@ -27,3 +35,9 @@ private:
 };
 
 #endif // VARIABLE_HPP
+
+
+
+Variable x_var;
+x_var.set_name('x')
+x_var.set_type(integer)
